==> data_analysis/model_draft/churn_rate_and_NBA/4_apply_trained_cox_model.ipynb <==
import pandas as pd
from lifelines import CoxPHFitter

import pickle
with open('data/learned_cox_model.pkl','rb') as f:
  cf = pickle.load(f)

data = pd.read_csv("data/merged_dataset_python.csv", index_col=0)

data["is_maried"] = data["is_maried"].astype(int)

# Exclude all fields except mentioned here
usefields = ["is_maried", "age", "work_experience", "children_number", "has_mortgage",
            "loan_products_nii", "checking_account_balance", "deposit_balance", "money_market_balance",
            "investable_asset_indicator", "Sentiment", "Sentiment_days_passed", 
            "Day_Observed", "Event"]
data = data[usefields]

predict = cf.predict_survival_function(data)

res = pd.DataFrame()
for i in [30, 60, 90, 120, 150, 180, 210, 240, 270, 300, 330, 365]:
    row = predict[:i].tail(1)
    row.index = [i]
    res = pd.concat([res, row])

# Define Final churn rate as 
Final_churn = ((1 - res.loc[365]) > 0.4).astype(int)
Final_churn.name = "Predicted churn"
Actual_churn = data.loc[:,"Event"]
Actual_churn.name = "Actual churn"
res = res.append([Final_churn, Actual_churn])

print "Accuracy = {0}%".format(100.0*sum(Actual_churn == Final_churn)/len(Final_churn))

res.T.to_csv("data/churn_prediction.csv")
