==> data_analysis/model_draft/churn_rate_and_NBA/5_apply_trained_rft_NBA_model.ipynb <==
import pandas as pd
from sklearn.ensemble import RandomForestClassifier

import pickle
with open('data/learned_rft_NBA_model.pkl','rb') as f:
  rfc = pickle.load(f)

data = pd.read_csv("data/merged_dataset_python.csv", index_col=0)

data["is_maried"] = data["is_maried"].astype(int)

# Exclude all fields except mentioned here
usefields = ["is_maried", "age", "work_experience", "children_number", "has_mortgage",
            "loan_products_nii", "checking_account_balance", "deposit_balance", "money_market_balance",
            "investable_asset_indicator", "Sentiment", "Sentiment_days_passed"]
y = data["NBO_choosen"]
data = data[usefields]

predict = rfc.predict_proba(data)

y_predict = rfc.predict(data)
print "Accuracy = {0}%".format(100.0*sum(y == y_predict)/len(y))

predict = pd.DataFrame(predict*100.0)
predict.columns = ["Investment_Discount", "HI_Deposit", "HI_MM", "Premium_account", "Refuses_all"]
predict.index = predict.index + 1

predict = pd.concat([predict, y], axis=1)

predict.to_csv("data/NBA_prediction.csv")
